images/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
exceptiongroup==1.1.0
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-django==4.5.2
sqlparse==0.4.3
tomli==2.0.1
Pillow
djangorestframework
sorl-thumbnail